Fixed backtick replacement bug.
There was a problem in the code where the post-substitution html parsing was expecting a paragraph to be produced, and sometimes it wasn't.
In fixing this, I also discovered that in some cases the tag containing the backticks could be removed, so I fixed this.
Added a first test to ensure behaviour!